Move utility method in filter manager and add unit tests
