+++ jsonrpc/eth_blockchain_test.go
@@ -4,9 +4,9 @@ import (
 	"errors"
 	"fmt"
 	"math/big"
-	"strconv"
 	"testing"
 
+	"github.com/0xPolygon/polygon-edge/blockchain"
 	"github.com/0xPolygon/polygon-edge/helper/progress"
 	"github.com/0xPolygon/polygon-edge/state/runtime"
 	"github.com/0xPolygon/polygon-edge/types"
@@ -97,102 +97,6 @@ func TestEth_Block_GetBlockTransactionCountByNumber(t *testing.T) {
 	assert.NoError(t, err)
 	assert.NotNil(t, res, "expected to return block, but got nil")
 	assert.Equal(t, res, 10)
-}
-
-func TestEth_Block_GetLogs(t *testing.T) {
-	t.Parallel()
-
-	blockHash := types.StringToHash("1")
-
-	// Topics we're searching for
-	topic1 := types.StringToHash("4")
-	topic2 := types.StringToHash("5")
-	topic3 := types.StringToHash("6")
-
-	var topics = [][]types.Hash{{topic1}, {topic2}, {topic3}}
-
-	testTable := []struct {
-		name           string
-		query          *LogQuery
-		shouldFail     bool
-		expectedLength int
-	}{
-		{"Found matching logs, fromBlock < toBlock",
-			&LogQuery{
-				fromBlock: 1,
-				toBlock:   3,
-				Topics:    topics,
-			},
-			false, 3},
-		{"Found matching logs, fromBlock == toBlock",
-			&LogQuery{
-				fromBlock: 2,
-				toBlock:   2,
-				Topics:    topics,
-			},
-			false, 1},
-		{"Found matching logs, BlockHash present",
-			&LogQuery{
-				BlockHash: &blockHash,
-				Topics:    topics,
-			},
-			false, 1},
-		{"No logs found", &LogQuery{
-			fromBlock: 4,
-			toBlock:   5,
-			Topics:    topics,
-		}, false, 0},
-		{"Invalid block range", &LogQuery{
-			fromBlock: 10,
-			toBlock:   5,
-			Topics:    topics,
-		}, true, 0},
-	}
-
-	// setup test
-	store := &mockBlockStore{}
-	store.topics = []types.Hash{topic1, topic2, topic3}
-	store.setupLogs()
-
-	for i := 0; i < 5; i++ {
-		store.add(&types.Block{
-			Header: &types.Header{
-				Number: uint64(i),
-				Hash:   types.StringToHash(strconv.Itoa(i)),
-			},
-			Transactions: []*types.Transaction{
-				{
-					Value: big.NewInt(10),
-				},
-				{
-					Value: big.NewInt(11),
-				},
-				{
-					Value: big.NewInt(12),
-				},
-			},
-		})
-	}
-
-	eth := newTestEthEndpoint(store)
-
-	for _, testCase := range testTable {
-		testCase := testCase
-		t.Run(testCase.name, func(t *testing.T) {
-			t.Parallel()
-
-			foundLogs, logError := eth.GetLogs(testCase.query)
-
-			if logError != nil && !testCase.shouldFail {
-				// If there is an error and test isn't expected to fail
-				t.Fatalf("Error: %v", logError)
-			} else if !testCase.shouldFail {
-				assert.Lenf(t, foundLogs, testCase.expectedLength, "Invalid number of logs found")
-			} else {
-				assert.Nil(t, foundLogs, "Expected first return param to be nil")
-			}
-		})
-	}
 }
 
 func TestEth_GetTransactionByHash(t *testing.T) {
@@ -583,6 +487,10 @@ func (m *mockBlockStore) GetAvgGasPrice() *big.Int {
 
 func (m *mockBlockStore) ApplyTxn(header *types.Header, txn *types.Transaction) (*runtime.ExecutionResult, error) {
 	return &runtime.ExecutionResult{Err: m.ethCallError}, nil
+}
+
+func (m *mockBlockStore) SubscribeEvents() blockchain.Subscription {
+	return nil
 }
 
 func newTestBlock(number uint64, hash types.Hash) *types.Block {
