+++ jsonrpc/filter_manager.go
@@ -216,6 +216,9 @@ type filterManagerStore interface {
 
 	// GetBlockByHash returns the block using the block hash
 	GetBlockByHash(hash types.Hash, full bool) (*types.Block, bool)
+
+	// GetBlockByNumber returns a block using the provided number
+	GetBlockByNumber(num uint64, full bool) (*types.Block, bool)
 }
 
 // FilterManager manages all running filters
@@ -344,6 +347,96 @@ func (f *FilterManager) Exists(id string) bool {
 	_, ok := f.filters[id]
 
 	return ok
+}
+
+func (f *FilterManager) GetLogsForQuery(query *LogQuery) ([]*Log, error) {
+	result := make([]*Log, 0)
+	parseReceipts := func(block *types.Block) error {
+		receipts, err := f.store.GetReceiptsByHash(block.Header.Hash)
+		if err != nil {
+			return err
+		}
+
+		for indx, receipt := range receipts {
+			for logIndx, log := range receipt.Logs {
+				if query.Match(log) {
+					result = append(result, &Log{
+						Address:     log.Address,
+						Topics:      log.Topics,
+						Data:        argBytes(log.Data),
+						BlockNumber: argUint64(block.Header.Number),
+						BlockHash:   block.Header.Hash,
+						TxHash:      block.Transactions[indx].Hash,
+						TxIndex:     argUint64(indx),
+						LogIndex:    argUint64(logIndx),
+					})
+				}
+			}
+		}
+
+		return nil
+	}
+
+	if query.BlockHash != nil {
+		block, ok := f.store.GetBlockByHash(*query.BlockHash, true)
+		if !ok {
+			return nil, fmt.Errorf("not found")
+		}
+
+		if len(block.Transactions) == 0 {
+			// no txs in block, return empty response
+			return result, nil
+		}
+
+		if err := parseReceipts(block); err != nil {
+			return nil, err
+		}
+
+		return result, nil
+	}
+
+	head := f.store.Header().Number
+
+	resolveNum := func(num BlockNumber) uint64 {
+		if num == PendingBlockNumber {
+			num = LatestBlockNumber
+		}
+
+		if num == EarliestBlockNumber {
+			num = 0
+		}
+
+		if num == LatestBlockNumber {
+			return head
+		}
+
+		return uint64(num)
+	}
+
+	from := resolveNum(query.fromBlock)
+	to := resolveNum(query.toBlock)
+
+	if to < from {
+		return nil, fmt.Errorf("incorrect range")
+	}
+
+	for i := from; i <= to; i++ {
+		block, ok := f.store.GetBlockByNumber(i, true)
+		if !ok {
+			break
+		}
+
+		if block.Header.Number == 0 || len(block.Transactions) == 0 {
+			// do not check logs in genesis and skip if no txs
+			continue
+		}
+
+		if err := parseReceipts(block); err != nil {
+			return nil, err
+		}
+	}
+
+	return result, nil
 }
 
 func (f *FilterManager) GetLogFilterFromID(id string) (*logFilter, error) {
